backend/requirements.txt
fastapi
uvicorn
motor
python-dotenv
passlib[argon2]
python-jose
pydantic
python-multipart
